requirements.txt
binance-futures-connector==4.0.0
python-dotenv==1.0.1